blblm {blblm}	R Documentation

Linear Regression with Little Bag of Bootstraps

Description

Linear or Generalized Linear Regression with Little Bag of Bootstraps.

Usage


blblm(formula, data, m = 10, B = 5000, parallel = FALSE)

Arguments

formula
a formula of model.

data
a data frame.

m
data splits, default = 10.

B
bootstraps, default = 5000.

parallel
logical value indicate use parallelization or not, default is not use.

Details

Linear Regression with Little Bag of Bootstraps

Value

res A list including estimates and sigma.

Author(s)

Maintainer: Zhanglin Shanggguan zshangguan@ucdavis.edu (ORCID)

[Package blblm version 0.1.0 Index]